Fixed crypteSeq implementation
- Moved char sequence manipulation in char_seq.h and created dedicated type

To validate the AppoLab exercice, run the following command: `./crypteSeqIO < planb_msg.txt | ./appolab`. Explanation: the first command (./crypteSeqIO < planb_msg.txt) will encrypt the contents of planb_msg.txt, then the output will be redirected as standard input to `./appolab` using the pipe operator. The second command will connect to the AppoLab server and send the encrypted content of the standard input.

This process is a bit cumbersome, and needs to be refactored before the project submission. We may merge crypteSeqIO and appolab to replicate this process automatically. The final file should be named crypteSeq.c (as specified in the submission requirements).